complete exercise 04_removeFromArray
this one gave me some trouble not gonna lie. I was held up by using array.length
for the loop. Everytime i removed an element it would change that value
which would make the next looping iteration look at the wrong index.
I tried multiple different ideas to get past this like instead of removing from the
inputted array I just added elements that didnt match the arguments into a new array
, this didnt work the way i though.
What did work though is simply reversing the direction in which i was iterating through
the array. Instead of working my way up and worked my way down.
This works because we dont care about the previous elements after we delete them and
once we delete a record we are moving on to the next one. EX: while goin up if I
got to the last element but had already deleted something i would be past that and wouldnt look at it
but now going down there is always going to be a 0th element if the array
isnt empty. Sorry if this is a lot, just proud i eventually got it, i know this isnt really what
commit messages are for.